+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 [< Repo](docs/repo.md) | [Index](README.md)
 <!-- nav:header:end -->
 
-<!-- last-updated-against: c08af8cc9db3b87f4c5a466915060f401e74fd7c -->
+<!-- last-updated-against: e2b82fc91941586d14fd8356ba6b2257ff824ac4 -->
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),